+++ changelog.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 - Removed support for billing-api.
+- Support for `create_customer` method removed in agent-api v3.6 and later.
 
 ## [0.4.1] - 2025-04-23